feat: incremental highlighting
Previously, there were no highlights-as-you-type, you'd see yaml string
highlighting until you close the expression with `}}`. If it was
a quoted string, you'd need to complete that too `}}"`.

The parser was designed to do it, but yaml injection query needs to be
adjusted:
* Inject when node has ${{ but don't require the finishing }}.
* Also look for ${{ in ERROR nodes of the yaml parser.

This change makes no modifications to the parser, just adds a missing
test and tweaks the suggested yaml injection query in the README.